projeto_leis/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0